PTU300/requirements.txt
pyserial==3.4


